Split up SignalGenerator and add Filter trait
The parameter and state handling is now done through a new trait `Equipment`. A
new `Filter` provides an interface for audio equipment that processes existing
signals instead of producing new ones. The previous implementation of filters
as a signal generator containing an underlying signal generator turned out to be
to limiting.